Tweak opportunity attack energy.
Don't randomize movement energy for a monster that just made an opportunity
attack, to avoid a situation in which the player moves away from a monster
and gets attacked twice that turn. Also, add a 50% chance that an
opportunity attack uses 1 energy, to be kinder to a player when an orc
warrior or similar catches up via random movement energy.