Make Jeremiah's gear more thematic
Instead of a glaive, give them a trident and a spooky magical orb.
What secrets lurk within? Perhaps it's occasionally an exciting
reward for the player?

This is intended as a nerf, since tridents usually do less damage
than glaives. A guaranteed glaive is really too much to give out
in mid-D, anyway.