Change Hexslinger starting spells
The new spell list is:
 -Jinxbite
 -Sigil of Binding
 -Inner Flame
 -Cause Fear
 -Dimensional Bullseye

Hexslinger was always troubled. Its spellbook was split across multiple
schools, making them tricky to cast, and they still didn't do much without
ALSO training ranged weapons heavily. Slow was often a trap and in
general the archetype was clunky, without providing sufficient benefit for
splitting your skills between weapons and spells.

The intent of this change is to provide a smoother curve of castable
spells, and more immediate value to training hexes instead of simply
ignoring them in favor of being a 'worse hunter'.

Jinxbite provides a little immediate power to help deal with earlygame
threats with which current hexslinger often struggles significantly.

Sigil of Binding provides powerful utility and can help set up Inner Flame
triggers in a way that the old spellset could almost never safely do in
practice.

Portal Projectile was always a spell that was far better late than it was
early, and so Dimensional Bullseye is now a capstone that provides
even stronger long-term value, but with other easier to cast spells that
let you bridge the gap to it better.

Cause Fear was always good (if you could manage to survive long enough
to have access to it!) and remains untouched.

(Also give them 1 starting level of Fire Magic, to help just a little bit
with getting Inner Flame off the ground)