Add Sigil of Binding: new level 3 Hexes spell
This spell creates two sigils on the ground at random locations within 2 tiles of
the caster (one at range 1 and the other at range 2). If a creature moved onto a
a sigil before it expires, that enemy will be bound in place for several turns.
This blocks all movement, but does *not* impair action, evasion, or otherwise
make the target more stabbable. When the binding expires, that enemy will become
Swift for twice as long as it was bound. Swift enemies are immune to sigils.

The duration of the binding scales with spellpower, but is also reduced by the
HD of the enemy. Only one cast worth of sigils can exist at a time (recasting
will replace them with new ones).

The caster is still bound in place if they step on their own sigil, so take care
about using it in confined spaces! (Self-binding does *not* scale with spellpower,
but the caster also does not become swift afterward)