Replace (player) Portal Projectile with Dimensional Bullseye
Portal Projectile was always in a slightly odd space. Its utility was
theoretically quite high, but also fairly situational. It only did anything for
one particular archetype, and despite being a low-ish level spell, usually not
very useful until midgame (since it also required strong ranged attack to do
anything reliable, and earlygame battles were less likely to present targets it
was relevant for). Hexslinger was clearly the class that *wanted* to get the
most use out of it, but didn't get very much out of pushing for it early.

Dimensional Bullseye takes the idea of 'smite-targeted ranged weapons', nudges
its level up to 4, and aims to make it more immediately impactful.

Now you smite-target an enemy with this spell to create a 'link' beween your
ranged weapon and it. And every time you perform a ranged attack aimed at any
*other* target (that isn't friendly or firewood), your projectile will teleport
to the marked target at the end of the attack, allowing you to hit 2 things at
once (with a spellpower-based accuracy bonus against the marked target).

It still consumes 1 mp per teleporting shot (this may or may not still be
necessary, since it can kill at most one target per cast anyway) and now works
with throwing weapons as well as launchers (the throwing ammo will land on the
ground at the marked target).

I've also made a small tweak to autofight that should hopefully make it
disfavor attacking a marked target if an unmarked target is also attackable (so
that this doesn't start asking people to manually aim in order to avoid wasting
the value of the spell)