Simplify Siphon Essence
Death Form's Siphon Essence was originally intended to provide an
exciting payoff for using the form in extended, where many enemies
are nonliving and hence immune to torment. However, Death Form's
resistances are already very exciting in extended, and it was
surprising and counterintuitive to players that Siphon Essence's
torment effect applied to non-tormentable foes.

Make the effect match Torment by only applying to living foes. For
now, the melee 'blight' Slow and Weak status effects can still apply
to non-nonliving enemies, but that may change.