Shift shapeshifting apts +1
It felt very bad to have the default apt for the skill be -2.
Shift it to -1, and adjust talisman skill levels such that each
talisman costs roughly the same XP to get to min and max skill on.

It now takes ~1.5x the XP to max out the top-level forms, instead of
~1.8x. Setting the apt to a default of +0 would mean taking only about
~1.25x the XP of getting a top-level form online to max it out; that
feels like too tight a squeeze to be a meaningful strategic choice.

Nerf Storm Form somewhat, since it's now easier to max out. Death Form
didn't really scale with power anyway (though we should fix that!), so
don't nerf it.