Experimental spell: Jinxbite (level 2 Hexes)
This spell gives a short self-buff to the player. Whenever they attack an
enemy while this buff is active, if that enemy fails a will check, you
spawn a jinxsprite near them. These behave similarly to foxfire, but when
reaching their target inflict a very minor amount of irresistable damage
and apply either slow or weak on a coinflip (along with the potential for
some very flavorful messages).

The spell can ONLY be cast while a susceptible enemy is in sight and will
expire immediately when one isn't (even if this was just from you turning
a corner). The idea here is to always create a minor opportunity cost to
casting the spell - earlygame the extra benefits are clearly worth it, but
should fall off later where you'd be better off just shooting something.

Behind the scenes, the base duration of jinxbite is fairly long, but each
jinxsprite you summon additionally reduces the duration of the buff on
you by several turns. And if a casting of jinxbite has not yet summoned a
single sprite, you have an increased chance beating the target's will.
This isn't something a player should need to concern themselves with; the
aim is just to make getting SOMETHING more consistent, with less bad low
rolls.