Experimental spell: Jinxbite (level 2 Hexes)
This spell gives a short self-buff to the player. Whenever they attack an
enemy while this buff is active, if that enemy fails a will check, you
inflict minor irresistable damage and draining (and some playful messages).
I realize draining is commonly necromantic, but several player abilities
give clear precedent for it being a source-agnostic (and the number of
minor debuffs we can hand out to enemies is a little limited...)

The spell can only be cast while a susceptible enemy is in sight and will
expire two turns after one no longer is (even if this was just from you
turning a corner to wait). Additionally, the player will suffer WILL/2
proportional to the duration of the spell when it is cast. The idea was to
create a minor opportunity cost to casting the spell - earlygame the extra
benefits are clearly worth it, but should fall off later where you'd be
better off just shooting something.

Behind the scenes, the base duration of jinxbite is fairly long, but each
time it triggers additionally reduces the duration of the buff on you by
4 turns (helping to make the number of triggers you get out of the spell
a little more consistent)