fix: prevent infinite loops in `dgn_place_item`
Previously, if `_apply_item_props` deterministically aborted, this loop
would not terminate. After 5a299da95614 this could trigger on
`grunt_nemelex_the_gamble`.  This commit adds a loop breaking condition
based on placement attempts; it's a bit heuristic but fixes this case,
and I don't see how it'd cause other problems. (It's unclear to me if
there was another way to trip the infinite looping, maybe not?)