fix: prevent unobtainable misc items altogether
This is a lot of engineering for exactly one vault, but failing to place
a misc item as the superb item in `grunt_nemelex_the_gamble` makes the
vault altogether pointless in most cases. So, prevent a superb item from
having misc type if it is flagged as unobtainable. (The other case to
worry about is unrands, but I believe these should reroll fine -- the
object base type doesn't get rerolled.)