Make Inner Flame irresistable, downscale damage and clouds based on power
Inner Flame was a powerful effect when everything worked, but it was often
very finicky in practice. Even when the hex landed, early characters would
frequently struggle to detonate the creature in a timely manner (and then
either fail to get almost any value out of it OR blow themselves up by
trying). I don't think that making it always land will make it too strong,
since you still need other, less-reliable tools to set it off.

But this commit also makes the POWER of the explosion scale with spellpower,
now that success chance does not. In practice, this means that spell-based
inner flame will usually do less damage and produce less clouds than
before (but the former may also make it more usable early on!) but
hopefully still feel better to use - especially if trying to set up
chain detonations.

Scroll of immolation should still result in detonation effects identical
to before this commit.