Various jinxbite changes
This spell no longer creates visible jinxsprites on activation, merely
directly inflicts damage+debuff on the affected enemies that fail a WL
check. There is no longer a bonus application chance if a single casting
hasn't triggered once yet and the base chance is lowered (to compensate
for that bug of the WL check being inverted... >.>).

Now instead of a coinflip of slow/weak, it applies drain instead. I was...
a little uncertain about this, given that most uses of draining are
necromantic in nature, but there's a shortage of minor debuffs to hand out
and Hep's Transference ability already gives some precedent.

There is now also a two-turn grace period of them losing interest when
there are no more enemies in sight, and the check happens AFTER enemy
movement (which should keep the spirit of the idea, while making it less
easy to randomly lose)

Finally, I am experimenting with the spell giving you 1/2 WILL proportional
to the duration of the buff, as a downside that ought to be less relevant
earlygame than it is later on.

(This doesn't remove the Jinxsprite creature itself yet; I think there may
be something else that can be done with them later)