Changes to Sigil of Binding
As a preventative measure for using this spell for slow, repeated kiting,
once binding wears off of an enemy, that enemy will gain swiftness for
twice as many turns as it was bound. While swift, enemies cannot be bound
again (and will harmlessly walk over future sigils). This should hopefully
make the spell useful as a temporary measure without making it overly
'safe' to use in a slow, tedious manner.

Additionally, letting the player dodge their own sigils while flying was
problematic for species with natural flight (and I don't want to make the
player swift either), so instead the player now suffers 3 mp loss whenever
they move onto a sigil (flying or not) and removes the sigil.