Brom's Barrelling Boulder (L4 Earth/Conj)
This spell creates a boulder in an adjacent tile to the player, which then
moves in a straight line in that direction. If the boulder tries to move
into a creature, it will damage it (and step into its tile if it dies in
the process). If the creature survives, it will be pushed back *and so will
every creature adjacent to it in a line*, effectively bowling an entire
line of enemies backward.

The boulder is a regular 'creature' that can be damaged, and takes
self-damage whenever it hits a creature (so that it will eventually die
even in the process of rolling over things). If it hits a wall, it will
immediately explode into a minor sharpnel explosion. (Like iood, it also
disappears if it leaves the player's LoS)

The spell is fun to play with and feels significantly different than iood
in its uses.

(Currently level 4 Conjurations/Earth. I sort of wanted it to be pure
Earth, but fear it would be too accessible at 4, and not able to fit into
the starter book at 5 (and then it's competing against LRD at that level
anyway....))

[Committer's note: refactored, bugfixed, and rebalanced. Very concerned
that this spell is wildly overpowered, but not being castable in melee
is a real restriction... seems fun to try out, at least.

Also added this to the EE start book, as seemed to be intended from the
commit message.]