Rending Blade: New level 4 Reaver starting spell
The Reaver kit was okay as-is, but with one of their 3 unique spells
being slated to get shared, there was some desire to give them something
else distinctive and here it is.

Aiming to fit Reaver's theme of 'conjurations that support melee and come
with situational downsides', this spell targets something in melee range,
and the two adjacent tiles, and then if at there are at least 2 enemies in
that range, spawns a Rending Blade nearby. (If there are 3 enemies, it
spawns 2). Each time the caster lands a hit in melee in the next few turns,
the blades will dance through a random line of hostile targets, dealing
moderate magic damage. Melee speed and accuracy directly affect its
usefulness, even if the damage comes purely from spellpower.

However, it comes with the strong tradeoff that the caster loses all
ability to cast spells while the blades are active! Again: something you
don't want to use every battle and comes with tradeoffs when you do, but
is a powerful effect for a hybrid conjurer.

I... wasn't sure exactly what schools to put it in. Level 4 pure
conjurations is a possibility, but felt too easy to access. Hexes was
another thought (given prism's position and the fact that this strongly
debuffs the caster), but I wondered if another inroad to Irradiate
might be appealing here and the spell *is* transforming their own spells
into tangible form... Open to changing this.