Buff Petrify a little
Reduce range slightly, but make it chain to up to 2 adjacent/connected
hostile targets at 2/3 enchantment power.

Player Petrify has been in an awkward state for a little while. It was the
capstone spell of the Earth Elementalist starting book for quite a while,
but was common for EEs to never make use it. Unreliably locking an enemy
down this way has questionable utility for a blaster playstyle with the
damage reduction involved. And while it can do decent work specifically as
a stabber setup spell, it's unclear that it performs *enough* better from
classic Enchanter spells to justify it being much harder to access due to
non-overlapping schools. Back when EE started with LRD, there was obvious
direct synergy (with being able to use AC-piercing LRD on a key target or
blow up groups in absence of walls), but that ship sailed a while ago now.

The hope here is that this will make the spell more interesting/usable as
disruption tool for non-stabbers as well, without making it significantly
better for stabbing individual key threats.

(While I'd only intended to modify the player spell (with a slight
renaming) and leave monster petrification unchanged, PleasingFungus has
argued that it's simpler if they both work the same way and that the
relatively minimal buff to the enemy spell is harmless. It's... probably
fine? Though I am sure some people will be caught off-guard by getting
petrified despite standing behind an ally...)