Petrify -> Petrifying Grasp (for players)
Reduce range slightly, make it chain to up to 2 adjacent/connected targets
at 2/3 power on cast.

Player Petrify has been in an awkward state for a little while. It is the
'capstone' spell of the Earth Elementalist starting book (ie: the highest
level), but it's much more common for EEs to never make use of it than
otherwise. Unreliably locking an enemy down this way has questionable
utility for a blaster playstyle with the damage reduction involved. And
while it can do decent work specifically as a stabber setup spell, it's
unclear that it performs *enough* better from classic Enchanter spells to
justify it being much harder to access due to schools. Back when EE started
with LRD, there was obvious direct synergy (with being able to use
AC-piercing LRD on a key target or blow up groups in absence of walls), but
that ship sailed a while ago now.

The hope here is that this will make the spell more interesting/usable as
disruption tool for non-stabbers as well, without making it significantly
better for stabbing key threats.