Standardize Passwall range at 3, move to level 3, give AC while using
The variable power-based range on Passwall was awkward in a few ways,
and this should make it a little less complex and more intuitive.

Also grant the player an AC bonus while transiting with the spell, to
make it feel a little safer and more tempting to use as an early-game
escape, when the margins of life and death are thinner.