Fix Inner Flame crash with starcursed masses and future-proof
Inner flame power wasn't being properly transferred on split (this would
affect slime creatures also).

I now also prevent the game from crashing on an Inner Flame explosion with
a missing power key, and just assume a power of 100 (ie: the pre-scaling
value) since this could be a problem for save upgrades from a version
before Inner Flame scaling.

It does print a diagnostic message saying that this probably shouldn't have
happened, in the hope of catching any future actual problems, but I am not
sure this is a good idea. It definitely can't assert on it, since there are
benign cases that will be impossible to separate in code from buggy ones.