feat: tweak welcome messages for non-standard modes
This always shows game mode explicitly with the welcome message (except
in normal, hints, and tutorial mode); previously it was only apparent
from this message for seeded mode. As part of this I changed the game
type name for descent mode to match the main menu, which hopefully won't
matter for anything (it would matter for database stuff, but aside from
welcome messages, there doesn't appear to be any).