Prevent a dragon armour trove placing for Draconians (#3272)
Unlike most other armour troves (which specifically place usable/aux
armour), this trove is themed around dragon armour and so is mostly useless
for Draconians. While there are a few usable items, it's already banned for
Octopodes (who would be able to use the same set of items from it).

Closes #3272.