feat: Add a more convenient exit to a WizLab (elliptic)
For Wucad Mu's Monastery, put the exit close to the starting area, since
we try to do this in general for portals. Previously the player had to
walk pretty far across the map to get to it and may have missed that it
even existed. Tweak the layout a bit, removing a room that's no longer
used and landscape trees appropriately now that the room is gone.