feat: Armataur rework: Rampage healing
Players and devs alike have expressed that Armataurs felt like they needed
something more, so here it is! This commit is a minor rework of At.

The new feature here is rampage healing: At XL 7, Armataurs now gain a
stacking regen buff when they roll towards enemies. This regen buff is
loosely based off of Powered by Death, but instead of triggering on kills,
it triggers when you roll.

Rampage healing only triggers when rolling towards threatening mons, as an
attempt to limit the long-term incentive to kite slow mons around.
This is an imperfect solution and I don't like how this can be opaque to
players; expect further simplifications in future.

Awkward Tongue is gone. Poof. Buh bye. It was awkward.

At's rugged brown scales 3 are replaced with tough skin 3: No more 7% MHP,
but you keep the 3 AC.

There are many different directions rampage healing could go from here, but
let's try it out this way for a bit?

Co-authored-by: Implojin <implosionwow@gmail.com>