Fix Eruption/Upheaval's tracer thinking it was radius 2 instead of 1.
Given that both Salamander Tyrants and Draconian Stormcallers come with
bands, this made them unwilling to use it in an enormous number of
situations where friendly fire wasn't even possible. (For instance, even
in an otherwise empty room, Tyrants could never use it on a turn the
player steps away from them in melee - resulting in a chase where they will
NEVER stop to cast their big spell).

As far as I can tell, this was broken since first implemention.