Make Eruption lava creation slightly more consistent
Given that Pyroclastic Surge (in most situations) specifically depends on
the combo with Eruption, it was a little weird that Eruption could
sometimes not create any lava whatsoever.

Instead of an independent coinflip per floor tile in its radius, it now
guarantees placing at least 1 lava tile per cast, if possible. (Still has
a coinflip for the rest of them)

These two commits are effectively a meaningful buff to Tyrants. If this
makes them too strong, their numbers can be nudged down somewhat, but I
think it's better for the player not to have invisible immunity to their
spells in ways that aren't transparent at all.