Prevent 'Done Exploring' when the player is fenced in by binding sigils
And instead properly show 'Partly explored, can't reach some places.'

Some parts of the pathfider code are a mess and needlessly hard to follow.
The description for this function in particular had clearly become untrue
in the 17 years since it was written, and I'm not even 100% certain this is
all it is used for anymore. Maybe it could use a better name?

I've tried to edit the comment to be a *bit* less deceptive, though, since
it clearly affects exploration pathfinding.