fix: tags.cc fixup for monster tile prop type change
cb44df74be1 changed the type of this prop from short to int, which led
to save compat crashes. be3b949103aec has a temporary fix at the point
of the crash. This commit moves the fix to tags.cc so that it happens on
load, and can be more easily kept with other save compat code.