Add unrand Charlatan's Orb
Adds a new unrand which boosts evocations, but delays teleports and
reduces stealth in the same way as the Orb of Zot.

This unrand serves two purposes besides thematic appeal; firstly it
provides a preview of the less obvious effects of the Orb status, and
secondly provides an equipment based means of boosting evocables.

[Committer's notes: Fixed punctuation and formatting, and did
checkwhite.]