Make Ozo's Fridge ignore firewood for huddling
Ozocubu's Refridgeration deals reduced damage to actors that are adjacent to another actor. This previously included firewood.

If fridge is your primary damage source (by a sufficiently large margin), you are encouraged to shoot firewood with a sling then wait for it to die when it comes into LOS, so that monsters could not huddle against it. This is an extremely unfun play pattern.

While this issue occurs to some degree with plant blocked missiles, it is not as severe in those cases, nor as easily fixable. In the case of missiles, retreating until an enemy enters line of fire is sufficient. In the case of Ozocubu's, you must retreat until the enemy is a full tile beyond the plant, which may be several turns where the enemy has LOF but can still huddle.