Make Chei piety gain only depend on monster base speed (hellmonk).
Previously players could gain more Chei piety while killing the same
monsters by going into statueform, waiting for the monster to haste itself,
not using abilities that slow the monster, and so on. This was fairly
insignificant in practice, but still a concern.

This commit simplifies Chei piety gain to only care about the monster's base
speed (and not any modifiers or the player speed). Instead of only giving
piety for monsters faster than the player, Chei only gives piety for
monsters who are speed 10 or faster. This has the largest impact on
spriggans, who will now gain piety from regular speed monsters as soon as
they worship Chei instead of having to find fast monsters to kill to gain
piety at the start.

The dependence on player speed was relatively small for non-spriggans in the
old system; a human worshipping Chei and killing a speed 10 monster
progressed from a 1+3/12 piety multiplier at start to a 1+5/12 piety
multiplier at max slowness. This commit replaces that with a global 1+4/12
piety multiplier. This commit also increases the bonus for killing fast
monsters, to compensate for no longer giving piety for slow monsters who
are still faster than the Chei-slowed player.