Species selection screen: DE -> Simple, Gh -> Intermediate, lower At.
DE is very simple (no mutations!) and has always been a popular choice for
a first casty character. Gr and Dr make fine casters too but they are less
obviously casting-devoted, so having a simple casting option in Simple
seems good.

Gh is reasonably simple if you play it as a heavy-armour Tr, but the low
Int/Dex can make it sort of miserable if you deviate much from that without
knowing exactly what you are doing. Having it next to Tr seems not great.

At was moved closer to the end of Simple because although it's generally
considered strong, it does have a solid number of weirdnesses (double
potions, weird body armour valuation, rampage).