Tweak the launcher loot in wizlab_tukima.
Make the (potential) randarts a mix of penetration and speed instead of
just speed, and don't generate a non-artefact triple crossbow of penetration
(since generating an artefact instead). This wizlab still has a high chance
of generating the best launcher in your game, but a much lower chance of
generating the best launcher in 100 games.