fix: a crash due to bad easy_floor_use code
The item use menu, for historical reasons, does some of its logic
outside of the menu code proper. This commit fixes a crash that resulted
from an ad hoc use of a keycode on menu exit to handle the
easy_floor_use case, where that keycode could also be set by normal menu
code via processing of CMD_MENU_CYCLE_HEADERS.  Instead, just add a bool
to specifically handle this return case.

The preconditions for this crash are an empty menu, which possibly
should be prevented (see 6eb203de8ee), and if allowed, could stand to
have better ui. But an empty menu still shouldn't crash.