fix: don't truncate headings when calculating column layouts
This was a bit of code that was incomplete in the multicolumn layout
code, where it was intended that headings always occupy their own row.
The side effect of it being unimplemented was that even for regular
1-column menus, headings could get truncated if they were longer than
regular menu items.  (No current multicolumn menus use headings, so this
remains a bit under-tested.)

Resolves #3305

(cherry picked from commit c67433417a310c1ef6ecb0a9b3e18459d0292db4)