feat: mention (e)quip in tutorial 3
On the theory that the (e)quip menu is simpler for new players to
understand, this adjusts a few things in tutorial 3 so that they teach
this key. I didn't actually remove mention of the specific keys, so it
does risk even more information overload. The tutorial summary is also
expanded and refactored.