fix: unrand generation logic in the abyss
If an acquirement scroll was read in the abyss, the previous iteration
of this logic would allow unrands *only* if they were lost in the abyss,
which wasn't intended. I think it would also potentially break logic for
items marked as lost in the abyss when *not* in the abyss.