Ban self-targeting /quicksilver (elliptic)
It's fun to have multiple uses for items, both offensive and utility,
but using wands of quicksilver to cancel effects on the player was
entirely too powerful. Since there are so many charges of /qs than
there are potions of cancellation, it effectively made most status
effects irrelevant. It's good for games to feel different depending
on which items you get, but this went too far, unfortunately.

Purple dracs can still self-target for now.