Vault review: Branch entries
Most of this is nerfs to notable statistical outliers of the most dangerous,
and minor buffs to the absolutely weakest. It's not bad for a branch entry
vault to have some notable danger, but the higher it is, the more it
de-emphasizes the branch the vault spawned in and the more it emphasizes
specific spoiler knowledge or arbitrary surprise spikes of danger. For
example, autoexplore being bad around trees in Lair:2-3 if any noise was
made because one could travel into blind harpooning beside dual bog bodies /
hydrae, or random doors in Depths revealing three hellephants and five
hell hogs immediately on the player at once. Not ideal.

(This occasionally also just uses more strict NSUBST behaviour to reduce
slim chances for absurd numbers, and also makes some minor adjustments
for Depths and Hells having continued to change identity over time.)

Additional notes:
 * A slim few rarely-spawning entries rely on a singular extreme out-of-depth:
   like a fire crab guarding Lair, or a guardian serpent guarding Snake.
   For now, I'm just adding some terrain cover to mildly reduce the threat;
   further discussion of the state of these is probably warranted.
   Some others, like the oklob plant guarding lair and the mini-Vaults:5
   entrance get people far more for hubris than inherent chasing danger...
   so their inherent nature hasn't been touched.

 * No deep water in Spider entries (no clinging anymore, discourage poking).

 * No ORIENT: float in Snake entries since other lair branch entries aren't.

 * While a couple of the tier 4 and tier 3 demons could serve the occasional
   support functionality (af_corrode, af_barbs, malmutate), their presence
   in Depths entries to Pan was notably weak otherwise compared to
   how much the constant Hells entrances got. With this, the entry function
   now places no tier 4s, less tier 3s, and usually a second tier 2.

 * This also slips in some minor formatting style standardization-
   stuff like a standard NAME -> TAGS -> DEPTH -> ORIENT -> MONS -> ITEM ->
   KFEAT -> [randomization] -> feature tiles and colours order, or
   minimizing "nothing" in monspec and itemspec. Establishing a firm
   house style guide and enforcing it might be nice eventually...?