Deduplicate Oka gifts (elliptic)
Ban Oka from offering non-artefact weapons/armour of the same type
and ego, since it was quite depressing to be offered three bardiches
of protection or whatever.

This is an Op^Oka armour gift buff especially, since they have so few
slots that they're very likely to get gifted artefacts. That's fine,
though.