Transform a tiny amount of Elf:3 loot into thematic Elf:2 Hall of Blades loot
The 1/3 chance to actually get an appropriate or even tempting top-tier
weapon as the only real rewards from the Hall of Blades is kind of weak
relative to the high danger an open vault eating up half of the level filled
with dancing weapons presents. It's regularly better to not open the vault,
but as a guaranteed placement avoiding it is awkward on top of the meagre
rewards- Elf:3 hatches, teleports, and teleport traps are much worse,
with little compensation for such an otherwise dramatic nature.

As such, I'm buffing Blade's items rewards, while trying to adhere to its
focus and flavour. Each vault gets a 50% chance for a blade talisman, so
unarmed characters might actually care about it and may get something, plus
two scrolls of enchant weapon and either a third or a brand weapon scroll.
A third handpicked top-tier dancing weapon also spawns half the time
(making the chance to get a specific weapon go from ~33% to ~42%), and their
handpicked brands list replaces venom and draining with heavy and spectral.
(These odds might be adjusted with further debate.)

To balance out such additional items in the branch, the Elf end vault
loot has been very slightly decreased- there's two items per endvault now
just replaced with gold. (This sort of arrangement might warrant further
reducing / "transferring" to support another optional branch's lacking
rewards versus danger, in Crypt, but that will take a fair bit more
review and debate.)