fix: Colour some clear stone in a vault (elliptic)
For gammafunk_dig_for_victory, make all clear stone use the special
color currently used to hold the deep troll earth mages flanking the
inner path. This colouration was done to distinguish diggable from
non-diggable clear walls, especially in console, and there's no reason
not to apply it to the clear stone proceeding the loot area as well. We
do that and additionally use the colored stone for the outer walls as
well, so now diggable and non-diggable clear walls are always
distinguished by the same color.