Make Sif gifts less frequent at lower piety.
Where Okawaru or Trog had 0 chance of gifting right after reaching 5* piety,
Sif had about an 8% chance; this made it reasonably plausible to get a gift
and then drop back below 5* during the piety timeout. This commit
generally makes Sif gifts less frequent at piety less than 6* and more
frequent at piety more than 6*. This does mean that the expected time to
the first Sif gift is longer; if this feels undesirable, I would suggest
moving gifts to start at 4* piety instead of 5*.

I'm not sure whether this is a large enough change to make it reasonably
easy for players to get to 6* Sif piety before getting every spell in the
game, but if it does have that effect then I see it as a good thing. The
main intent of the change though is to make it much less likely to drop
a piety rank without using abilities.