feat: clo to force a full des cache reset
This option ensures that the des cache is rebuilt from scratch,
independent of whether the .des files have changed. It then uses this
option in `make test`, so that local runs of `make test` exercise the
entire set of des files (and match CI and webserver version update
behavior for cases where des syntax changes).

(For consistency, I probably should have named this `-resetdb`, but I
don't like the name `-builddb` in the first place.)