Remove weapons from some zombies in a bailey
Zombies aren't supposed to be able to use items, and giving them polearms
in this bailey results in the polearms not being visibly displayed in
console. The main threat of the vault is the gargoyles and vampires, so no
compensation has been given to the zombies (but they could perhaps instead
be changed back to regular orcs, as well).