New monster hex - Vitrify (take 30% more damage!)
Aside from the slow propagation of Barbs and Will/2, new statuses haven't
really been spread around on monsters too often. There's a large pool of
old monsters exchanging back and forth a variety of straightforward
conjurations and hexes while newer monsters sit on weird one-offs that are
risky to spread much further due to their given distinguishing complexity.
Surely something new could help freshen up such a pool and allow another
high-threat effect.

This new Vitrify hex applying the Fragile debuff has the same effect as the
downside of two stacks of Harm without the upside: if you take any damage,
you'll take 30% more. This is obviously quite bad, but in comparison to
paralysis, which this is meant to mildly lower the presence of, one still
keeps EV and SH to avoid damage with. Also, instead of rolling 2-4 turns of
monster behaviour to randomly fail to / ruin you with with little recourse,
one has 10-25 turns to use any number of defensive consumables, spells, and
actions to shield one's self with until it wears off. Currently, like Strip
Willpower, monsters won't cast it again if you already have the debuff, so
they can focus on hitting you instead.

(There's been some wondering if lowering current + max HP would be more
transparent about its exact effects in comparison to the innate fuzziness
and spikes inherent to screwing with damage in a game with such complicated
defense calculations; c.f. the time gap between 762e91d and 141e18f.
I'm not much of a fan of this alternative; aside from stepping on the toes
of the rather widespread flayed ghost, I'd argue that for many players the
actual value of one's own max HP numbers is pretty fuzzy in how most people
don't remember the exacts of vast network of incoming damage numbers anyway.
Still, it's worth noting for future work if this state of Vitrify doesn't
work out.)

(Balancing numbers, word and phrasing choices, and uses of this, are all
openly subject to change in the future. At the very least, it's been nine
years since the Vitrification card was around and seven since Xom did the
same effect, so it's unlikely for too much confusion on that front, and
while Fragile is also an equipment modifier word, I'm sure most people
would agree unequipping your self is lethal. This has also been implemented
as a status for monsters, for charming or summoning purposes. It's rather
boring to directly put much of in player hands, but it might be reasonable
to toss into Xom's list or shuffle around some Nemelex cards to fit in
over Degeneration's awkward placement in the destruction deck.)