Simplify Static Discharge damage calculations
For mysterious reasons, Static Discharge had 3 different damage
calculations - one for hitting monsters, one for the hitting the player, and
one for when the spell caused self-damage to the caster. However, the self-
damage calculation was always overwritten by the player-specific calculation
so only applied to monster self-damage.

Simplify the spell to always use the same damage calculation, which is
halved when hitting the caster. In practice this should be mostly
indistinguishable - the player will now take noticeably more damage from
hostile Static Discharge casters, but there aren't any of these outside of
player ghosts.