Apply half AC to Static Discharge damage
Instead of ignoring AC entirely, for consistency with other spells dealing
BEAM_ELECTRICITY damage. Increase base damage slightly to compensate
(although this may not be necessary, since most early monsters won't have
enough AC for it to make a difference, and as an L2 spell it should fall
off past the early game).