feat: Add weapon damage ratings to clua
The new damage_rating() function in the item metatable gives both the
rating as a number and the full damage rating string as return values.
An input argument added to the crawl-side damage_rating() function
allows us to get the first value. This might be doable in a cleaner way,
but this approach was simplest given how the calculation is structured.